requirements.txt
requests
bs4
supabase
python-dotenv
finance-datareader
plotly
pandas